+++ src/app/locales/en/tabs/spaces.js
@@ -2,6 +2,7 @@ export default {
   menuItem: 'Data',
   aspects: {
     index: 'Space overview',
+    data: 'Data',
     providers: 'Space providers',
     members: 'Space members',
     membership: 'Membership',