+++ src/app/services/content-resources.js
@@ -9,7 +9,7 @@
 
 import { inject as service } from '@ember/service';
 import Service from '@ember/service';
-import { reject } from 'rsvp';
+import { resolve, reject } from 'rsvp';
 
 export default Service.extend({
   providerManager: service(),
@@ -19,6 +19,7 @@ export default Service.extend({
   groupManager: service(),
   clusterManager: service(),
   harvesterManager: service(),
+  uploadManager: service(),
 
   /**
    * @param {string} type plural type of tab, eg. providers
@@ -42,6 +43,11 @@ export default Service.extend({
         return this.get('groupManager').getRecord(id);
       case 'harvesters':
         return this.get('harvesterManager').getRecord(id);
+      case 'uploads': {
+        const oneprovider = this.get('uploadManager.sidebarOneproviders')
+          .findBy('id', id);
+        return oneprovider ? resolve(oneprovider) : reject();
+      }
       default:
         return reject('No such model type: ' + type);
     }