+++ src/app/services/sidebar-resources.js
@@ -25,6 +25,7 @@ export default SidebarResources.extend({
   clusterManager: service(),
   harvesterManager: service(),
   harvesterActions: service(),
+  uploadManager: service(),
 
   /**
    * @param {string} type
@@ -44,6 +45,10 @@ export default SidebarResources.extend({
         return this.get('groupManager').getGroups();
       case 'harvesters':
         return this.get('harvesterManager').getHarvesters();
+      case 'uploads':
+        return resolve({
+          list: this.get('uploadManager.sidebarOneproviders'),
+        });
       case 'users':
         return this.get('currentUser').getCurrentUserRecord().then(user => {
           return resolve({ list: A([user]) });
@@ -78,9 +83,11 @@ export default SidebarResources.extend({
   /**
    * @override
    */
-  getItemsSortingFor(type) {
-    if (type === 'tokens') {
+  getItemsSortingFor(resourceType) {
+    if (resourceType === 'tokens') {
       return ['isActive:desc', 'name'];
+    } else if (resourceType === 'uploads') {
+      return ['isAllOneproviders:desc', 'name'];
     } else {
       return this._super(...arguments);
     }