+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5879 Changed provider permissions to support permissions in space
 * VFS-5782 Added simplified version of new tokens GUI
 * VFS-5020 Added Ceph aspect to Oneprovider cluster
 * VFS-5813 Added translations for new possible backend errors